requirements.txt
absl-py==2.1.0
Adv2==1.2.0
astropy==6.1.7
astropy-iers-data==0.2025.3.3.0.34.45
astroquery==0.4.9.post1
asttokens==3.0.0
attrs==25.1.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==4.4.2
ephem==4.2
exceptiongroup==1.2.2
executing==2.2.0
flatbuffers==25.2.10
fonttools==4.56.0
html5lib==1.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.33.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.2
JsonForm==0.0.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
JsonSir==0.0.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keyring==25.6.0
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.43.0
logging==0.4.9.6
matplotlib==3.10.1
matplotlib-inline==0.1.7
mediapipe==0.10.21
ml_dtypes==0.5.1
more-itertools==10.6.0
moviepy==1.0.3
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.60.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
proglog==0.1.10
prompt_toolkit==3.0.50
protobuf==4.25.6
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.22
pyephem==9.99
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.1
pypiwin32==223
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyqtgraph==0.12.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
Python-EasyConfig==0.1.7
pytz==2025.1
pyvo==1.6.1
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
pyzmq==26.2.1
ravf==1.0.1
referencing==0.36.2
requests==2.32.3
Resource==0.2.1
rpds-py==0.23.1
scikit-image==0.24.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
soupsieve==2.6
srt==3.5.3
stack-data==0.6.3
tifffile==2025.2.18
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vosk==0.3.45
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0
webvtt-py==0.5.1
winshell==0.6
zipp==3.21.0
